requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
scikit-learn==1.0.2
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1
joblib==1.1.0
email-validator==1.1.3
Werkzeug==2.0.3
google-generativeai==0.3.1
